fix: 명함 공유 시 스티커 배경색이 적용되지 않던 문제 해결

+++ app/src/main/java/com/hongul/filq/model/BusinessCard.serial.kt
@@ -1,5 +1,6 @@
 package com.hongul.filq.model
 
+import android.util.Log
 import androidx.compose.ui.graphics.Color
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -63,7 +64,7 @@ object ColorSerializer: KSerializer<Color> {
         PrimitiveSerialDescriptor("androidx.compose.ui.graphics.Color", PrimitiveKind.LONG)
 
     override fun deserialize(decoder: Decoder): Color {
-        return Color(decoder.decodeLong())
+        return Color(decoder.decodeLong().toULong())
     }
 
     override fun serialize(encoder: Encoder, value: Color) {
